Significantly improved html regeneration, though not yet in the normal update-a-backup case. If you run the tool with the new -g option, all html files will be regenerated from the xml backup data.
User pic data is still something of a mess in two hashes, but I've
started making it more sensible. The hashes are now attributes of the
source journal object, which is at least sane.

Also, we now download, store, and use custom group data. If an entry
is flocked or filtered in any way, we say so in the html version of
the entry. I'm probably not showing private entries properly yet.


git-svn-id: svn+ssh://kelphw.com/home/svn/ljmigrate@56 9f68e313-fd7d-4df9-bacf-b2c4cf5f96a2