Bomb-proof failures to fetch comments vi export_comments.bml.
These fetches use the now-deliberately-broken method of passing cookies
to auth, so they're never going to work this way. But this is a fast
way of allowing the script to finish its work gracefully.

Also, change the sample config to nudge people toward using
www.livejournal.com, which works. Stripping the www off doesn't, for
reasons I cannot guess at.