Adapted to Livejournal's infamous release 86 and made comment backup work again.
If the site we're backing up from is livejournal, we now post to
login.bml and scrape the session cookies from the response. We then use
those cookies to authenticate when fetching comments using
export_comments.bml. If we're backing up from a site other than
Livejournal itself, we use the older, simpler challenge/response call
to request a session cookie.