Add support for using plain-text password auth for dreamwidth
As explained here: https://dw-dev.dreamwidth.org/221167.html

Dreamwidth has stopped supporting the challenge-response auth mechanism
because this requires storing plain-text passwords on the server.

Instead dreamwidth now expects the usage of plain-text passwords,
relying on the TLS connection to the server to protect these.

Change ljmigrate to use plain-text password when connecting to
dreamwidth so that it can work with dreamwidth again.

This is loosely based on a similar pull-req for ljdump:
https://github.com/ghewgill/ljdump/pull/8